Move examples to folder and ignore in npm
