+++ examples/example4.js
@@ -5,7 +5,7 @@
  *
  */
 
-var onvif = require('./lib/onvif');
+var onvif = require('../lib/onvif');
 var xml2js = require('xml2js')
 var stripPrefix = require('xml2js').processors.stripPrefix;
 
@@ -17,7 +17,7 @@ onvif.Discovery.on('device', function(cam,rinfo,xml){
 	console.log(cam.hostname + ':' + cam.port + cam.path);
 })
 */
-onvif.Discovery.on('device', function (cam, rinfo, xml) {
+onvif.Discovery.on('device', function(cam, rinfo, xml) {
 	// Function will be called as soon as the NVT responses
 
 
@@ -32,8 +32,8 @@ onvif.Discovery.on('device', function (cam, rinfo, xml) {
 		tagNameProcessors: [stripPrefix]   // strip namespace eg tt:Data -> Data
 	});
 	parser.parseString(xml,
-		function (err, result) {
-			if (err) return;
+		function(err, result) {
+			if (err) {return;}
 			let urn = result['Envelope']['Body'][0]['ProbeMatches'][0]['ProbeMatch'][0]['EndpointReference'][0]['Address'][0].payload;
 			let xaddrs = result['Envelope']['Body'][0]['ProbeMatches'][0]['ProbeMatch'][0]['XAddrs'][0].payload;
 			let scopes = result['Envelope']['Body'][0]['ProbeMatches'][0]['ProbeMatch'][0]['Scopes'][0].payload;
@@ -42,15 +42,15 @@ onvif.Discovery.on('device', function (cam, rinfo, xml) {
 			let hardware = "";
 			let name = "";
 			for (let i = 0; i < scopes.length; i++) {
-				if (scopes[i].includes('onvif://www.onvif.org/name')) name = decodeURI(scopes[i].substring(27));
-				if (scopes[i].includes('onvif://www.onvif.org/hardware')) hardware = decodeURI(scopes[i].substring(31));
+				if (scopes[i].includes('onvif://www.onvif.org/name')) {name = decodeURI(scopes[i].substring(27));}
+				if (scopes[i].includes('onvif://www.onvif.org/hardware')) {hardware = decodeURI(scopes[i].substring(31));}
 			}
 			let msg = 'Discovery Reply from ' + rinfo.address + ' (' + name + ') (' + hardware + ') (' + xaddrs + ') (' + urn + ')';
 			console.log(msg);
 		}
 	);
 })
-onvif.Discovery.on('error', function (err,xml) {
+onvif.Discovery.on('error', function(err,xml) {
 	// The ONVIF library had problems parsing some XML
 	console.log('Discovery error ' + err);
 });
